# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Eneverre NVR API
  version: 1.0.0
  description: |
    API for camera management, streaming and playback.

    Typical flow:
    1. GET /api/cameras
    2. Pick a camera (playback=true if you want recordings)
    3. GET /api/camera/{id}/recordings/list
    4. GET /api/camera/{id}/recordings/get
servers:
  - url: http://localhost:8080

# Most endpoints accept either HTTP Basic or a Bearer access token.
security:
  - basicAuth: []
  - bearerAuth: []
paths:
...
  /api/app/tv/update:
    get:
      summary: Latest available build for the Android TV client
      description: |
        Returns the current manifest for the TV track, or `204 No Content`
        when no build has been published yet. Used by `EneverreTV` on cold
        start to check for an update. Anonymous (no auth). See `doc/UPDATES.md`
        for the full client protocol.
      security: []
      responses:
        '200':
          description: A build is available
          headers:
            Cache-Control:
              schema: { type: string, example: "no-store" }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateManifest'
        '204':
          description: No build has been published for this track
        '503':
          description: Auto-update is not configured on this server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/app/phone/update:
    get:
      summary: Latest available build for the Android phone client
      description: Same shape as the TV endpoint, separate track. Anonymous.
      security: []
      responses:
        '200':
          description: A build is available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateManifest'
        '204':
          description: No build has been published for this track
        '503':
          description: Auto-update is not configured on this server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      description: Access token issued by POST /api/auth/login (or the device flow).
    updateToken:
      type: http
      scheme: bearer
      description: |
        Publish-only bearer token matching the server's `[updates] publish_token`
        (or `ENEVERRE_UPDATES_PUBLISH_TOKEN`). The publish endpoints
        (POST /api/admin/app/updates/{tv,phone}) accept ONLY this token;
        user/password and session Bearer are rejected. If no token is
        configured on the server, the publish endpoints return 503.
...
  schemas:
...
    UpdateManifest:
      type: object
      description: |
        Body of `GET /api/app/<track>/update` when a build is available.
        Clients upgrade iff `versionCode` is strictly greater than their
        installed `BuildConfig.VERSION_CODE`. The `url` points to the
        currently advertised APK; `sha256` lets the client verify the
        download. `mandatory` removes the "Later" / "Skip" buttons in the
        dialog.
      required: [versionName, versionCode, apkFilename, sha256, url]
      properties:
        versionName:
          type: string
          example: "1.0.1"
        versionCode:
          type: integer
          example: 10101
        apkFilename:
          type: string
          description: Basename of the APK as stored on the server
          example: "eneverre-tv-universal-1.0.1.apk"
        url:
          type: string
          format: uri
          description: Absolute URL the client should GET to download the APK
        sha256:
          type: string
          description: Lowercase hex SHA-256 of the APK bytes
          example: "9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08"
        mandatory:
          type: boolean
          default: false
        releaseNotes:
          type: string
          description: Optional release notes shown in the update dialog
    Error:
      type: object
      description: Standard error body
      properties:
        detail:
          type: string
...